application/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-mongodb-backend==5.2.0b0
djangorestframework==3.16.0
dnspython==2.7.0
pymongo==4.12.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
